Do clean up get fn and move some logic to generate_otp
I used a naive approach to handle the case of wrong password for an encrypted code.
Basically it is based on Err return from crypt fn and const TALARIA.
There is a very low probability that someone is going to use the password as I took
for TALARIA. I moved this logic from get and ls to generate_otp fn.