Update all functions - by default encryption is enabled
I contemplated the concepts of encryption and simplicity, and I concluded
that the optimal approach is to make encryption mandatory while still
allowing users the option to deactivate it. Consequently, I implemented
this solution. As a result of my most recent code modifications,
encryption is now enabled by default. However, if a user includes
the -u or --unencrypt flag, the code will be stored without any encryption.